pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vectorize"
version = "0.1.0"
description = "CLI tool to convert raster images to SVG"
requires-python = ">=3.10"
dependencies = [
    "vtracer>=0.6.10",
    "Pillow>=10.0",
    "click>=8.1",
    "lxml>=5.0",
]

[project.scripts]
vectorize = "vectorize.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]
